API get and save

+++ src/main/kotlin/com/example/anymind/posgateway/model/PaymentDO.kt
@@ -8,7 +8,7 @@ data class PaymentDO(
     val finalPrice: Double,
     val points: Int,
     val customerId: String,
-    val requestedPrice: Int,
+    val requestedPrice: Double,
     val priceModifier: Double,
     val createdAt: Instant,
     val metadata: PaymentMethodMetadata,
